feat: 新增知识库管理模块，支持研报元数据导入与状态查询
- 新增后端知识库API路由、服务层、数据模型与Schema定义
- 新增前端知识库管理页面与对应API接口
- 优化异常处理与数据库会话管理
- 更新API文档与项目说明文档

+++ backendV2/app/models/income_sheet.py
@@ -46,42 +46,42 @@ class IncomeSheet(Base):
         nullable=False,
         comment="报告类型：REPORT/SUMMARY，由 financial_report 主表统一回填",
     )
-    net_profit = Column[DECIMAL](DECIMAL(20, 2), comment="净利润(万元)")
-    net_profit_yoy_growth = Column[DECIMAL](DECIMAL(10, 4), comment="净利润同比(%)")
-    other_income = Column[DECIMAL](DECIMAL(20, 2), comment="其他收益（万元）")
-    total_operating_revenue = Column[DECIMAL](
+    net_profit = Column[float](DECIMAL(20, 2), comment="净利润(万元)")
+    net_profit_yoy_growth = Column[float](DECIMAL(10, 4), comment="净利润同比(%)")
+    other_income = Column[float](DECIMAL(20, 2), comment="其他收益（万元）")
+    total_operating_revenue = Column[float](
         DECIMAL(20, 2), comment="营业总收入(万元)"
     )
-    operating_revenue_yoy_growth = Column[DECIMAL](
+    operating_revenue_yoy_growth = Column[float](
         DECIMAL(10, 4), comment="营业总收入同比(%)"
     )
-    operating_expense_cost_of_sales = Column[DECIMAL](
+    operating_expense_cost_of_sales = Column[float](
         DECIMAL(20, 2), comment="营业总支出-营业支出(万元)"
     )
-    operating_expense_selling_expenses = Column[DECIMAL](
+    operating_expense_selling_expenses = Column[float](
         DECIMAL(20, 2), comment="营业总支出-销售费用(万元)"
     )
-    operating_expense_administrative_expenses = Column[DECIMAL](
+    operating_expense_administrative_expenses = Column[float](
         DECIMAL(20, 2), comment="营业总支出-管理费用(万元)"
     )
-    operating_expense_financial_expenses = Column[DECIMAL](
+    operating_expense_financial_expenses = Column[float](
         DECIMAL(20, 2), comment="营业总支出-财务费用(万元)"
     )
-    operating_expense_rnd_expenses = Column[DECIMAL](
+    operating_expense_rnd_expenses = Column[float](
         DECIMAL(20, 2), comment="营业总支出-研发费用（万元）"
     )
-    operating_expense_taxes_and_surcharges = Column[DECIMAL](
+    operating_expense_taxes_and_surcharges = Column[float](
         DECIMAL(20, 2), comment="营业总支出-税金及附加（万元）"
     )
-    total_operating_expenses = Column[DECIMAL](
+    total_operating_expenses = Column[float](
         DECIMAL(20, 2), comment="营业总支出(万元)"
     )
-    operating_profit = Column[DECIMAL](DECIMAL(20, 2), comment="营业利润(万元)")
-    total_profit = Column[DECIMAL](DECIMAL(20, 2), comment="利润总额(万元)")
-    asset_impairment_loss = Column[DECIMAL](
+    operating_profit = Column[float](DECIMAL(20, 2), comment="营业利润(万元)")
+    total_profit = Column[float](DECIMAL(20, 2), comment="利润总额(万元)")
+    asset_impairment_loss = Column[float](
         DECIMAL(20, 2), comment="资产减值损失（万元）"
     )
-    credit_impairment_loss = Column[DECIMAL](
+    credit_impairment_loss = Column[float](
         DECIMAL(20, 2), comment="信用减值损失（万元）"
     )
